requirements.txt
streamlit
python-dotenv
google-genai
pandas
numpy
requests
pillow
streamlit-option-menu
streamlit>=1.0